# awa/docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AWA API",
    "description": "Agentic Workflow Accelerator API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/workflows/list": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Workflows: List all workflows",
        "description": "List all available workflows from both core application and external workers.",
        "operationId": "list_workflows_api_v1_workflows_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_queue",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by task queue",
              "title": "Task Queue"
            },
            "description": "Filter by task queue"
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/runs": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Workflow Runs: List workflow runs",
        "description": "List all running and completed workflow runs.",
        "operationId": "workflow_runs_api_v1_workflows_runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Workflows: Run workflow",
        "description": "Run a workflow.",
        "operationId": "run_workflow_api_v1_workflows_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRunPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Run Workflow Api V1 Workflows Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/runs/{run_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Workflows: Get workflow run details",
        "description": "Get detailed information about a specific workflow run by run ID.",
        "operationId": "get_workflow_run_api_v1_workflows_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{workflow_id}/agent-sessions": {
      "get": {
        "tags": [
          "workflows",
          "agent-streaming"
        ],
        "summary": "Get workflow agent sessions",
        "description": "Get all agent streaming sessions associated with a specific workflow.",
        "operationId": "get_workflow_agent_sessions_api_v1_workflows__workflow_id__agent_sessions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowAgentSessionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/tasks": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "HITL: List HITL tasks",
        "description": "List all Human-in-the-Loop tasks awaiting interaction.",
        "operationId": "list_hitl_tasks_api_v1_hitl_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HITLTaskInfo"
                  },
                  "title": "Response List Hitl Tasks Api V1 Hitl Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/tasks/{workflow_id}": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "HITL: List HITL tasks for workflow",
        "description": "List all HITL tasks for a specific parent workflow.",
        "operationId": "list_hitl_tasks_for_workflow_api_v1_hitl_tasks__workflow_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HITLTaskInfo"
                  },
                  "title": "Response List Hitl Tasks For Workflow Api V1 Hitl Tasks  Workflow Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/task/{workflow_id}": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "HITL: Get HITL task details",
        "description": "Get detailed information about a specific HITL task.",
        "operationId": "get_hitl_task_api_v1_hitl_task__workflow_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HITLTaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/task/{workflow_id}/submit": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "HITL: Submit HITL task response",
        "description": "Submit a response to complete a HITL task.",
        "operationId": "submit_hitl_task_api_v1_hitl_task__workflow_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Response"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Submit Hitl Task Api V1 Hitl Task  Workflow Id  Submit Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/task/{workflow_id}/message": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "HITL: Send HITL chat message",
        "description": "Send a chat message to a HITL task in chat mode.",
        "operationId": "send_hitl_message_api_v1_hitl_task__workflow_id__message_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Message Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Send Hitl Message Api V1 Hitl Task  Workflow Id  Message Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/task/{workflow_id}/chat/history": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "HITL: Get HITL chat history",
        "description": "Get the current chat history for a HITL task.",
        "operationId": "get_hitl_chat_history_api_v1_hitl_task__workflow_id__chat_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Hitl Chat History Api V1 Hitl Task  Workflow Id  Chat History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/task/{workflow_id}/chat/user-message": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "HITL: Send user message to HITL task",
        "description": "Send a user message to a HITL task, triggering workflow processing.",
        "operationId": "send_hitl_user_message_api_v1_hitl_task__workflow_id__chat_user_message_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "awa_auth_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Awa Auth Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Message Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Send Hitl User Message Api V1 Hitl Task  Workflow Id  Chat User Message Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workers/register": {
      "post": {
        "tags": [
          "workers"
        ],
        "summary": "Registration: Register worker capabilities",
        "description": "Register a worker with its available workflows and activities.\n\n#### Schema Requirements:\n* Must have `type: object` at root level\n* Must include `properties` field containing property definitions\n* Each property must specify a `type` field\n* Optional: Include `required` array for mandatory fields\n* Optional: Add validation rules (minimum, maximum, pattern, etc.)",
        "operationId": "register_worker_api_v1_workers_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerRegistrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agents/stream/emit": {
      "post": {
        "tags": [
          "agent-streaming"
        ],
        "summary": "Emit stream event",
        "description": "Emit a stream event via HTTP (stores and broadcasts via Socket.IO). Service-authenticated only.",
        "operationId": "emit_stream_event_api_v1_agents_stream_emit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WorkflowListResponse": {
        "properties": {
          "workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowInfo"
            },
            "type": "array",
            "title": "Workflows"
          }
        },
        "type": "object",
        "required": [
          "workflows"
        ],
        "title": "WorkflowListResponse",
        "description": "Response containing a list of available workflows.\n\nAttributes:\n    workflows: List of workflow information objects"
      },
...
      "WorkflowRunListResponse": {
        "properties": {
          "workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowRun"
            },
            "type": "array",
            "title": "Workflows"
          }
        },
        "type": "object",
        "required": [
          "workflows"
        ],
        "title": "WorkflowRunListResponse",
        "description": "Lists running and completed workflows.\n\nAttributes:\n    data: List of Workflow Runs"
      },
      "WorkflowRunPayload": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "input": {
            "type": "string",
            "title": "Input"
          },
          "task_queue": {
            "type": "string",
            "title": "Task Queue",
            "description": "Optional task queue for workflow execution"
          }
        },
        "type": "object",
        "required": [
          "name",
          "input"
        ],
        "title": "WorkflowRunPayload",
        "description": "POST data needed to initate a new workflow run.\n\nAttributes:\n    name: The name of the workflow to initiate\n    input: Input payload to initiate workflow, in json\n    task_queue: Optional task queue to use for workflow execution"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}